No need to repeat checks for IO::Socket::SSL in t/*_ssl.t
We have already skipped all tests if !can_ssl() so there is no need to
repeat what it checks. Nor is there any need to test if
IO::Socket::SSL::Utils can be loaded and contains a required function: We
can safely assume this is true if we've already checked that a suitable
version of IO::Socket::SSL is available. All we need to do is to actually
load IO::Socket::SSL::Utils since IO::Socket::SSL doesn't load it itself.