Increased minimum required Socket version from 1.3 to 2.016
This turned out to be the cause of the remaining problem on CPAN RT#100020,
but only on machines having IO::Socket::IP (and even then that worked fine
on my machine, so it is probably OS dependent).

Don't restate the minimum required version everywhere that Socket is used.
It seems like overkill (and maintenance hassle) in a world where various
CPAN shells do a good job of installing required versions of prerequisite
modules, and it already wasn't done for IO::Socket (which also has a minimum
required version).