last does not work in a do {} block
There is no point in continuing with the auth() function if there are no
mechanisms left to try (it would only issue an AUTH command with no
arguments, which is a syntax error) so simply return instead of trying to
jump out of the do {} block.

Also return if we cannot get the failed mechanism from the client because
in that case we have no way of changing the list of mechanisms to try.

Partially fixes CPAN RT#100235.