Optional automatic hello
Added option "SendHello => 0|1" to constructor Net::SMTP::new().
If an option "SendHello" is passed and set to 0 then command HELO will not be sent during calling of constructor Net::SMTP::new() and it has to be sent manually using method Net::SMTP::hello(). If the option isn't passed or set to 1 then the constructor works in the same manner as in earlier versions (command HELO is sent automatically in the constructor). The fact that option is on by default allows to keep compatibility with previous versions of the library. Performing Net::SMTP::hello() manually allows to handle the result of the command in the same way as for other commands.