Fix Net::Cmd::datasend() for octets stored in an upgraded string
The data passed to datasend() should already be encoded, but it can
sometimes happen that the string holding the octets gets accidentally
upgraded and it was wrong for datasend() to treat it differently in
that case.

Fixes CPAN RT#104433. Many thanks to Ricardo and Aristotle for their help
on the ticket.