Add a workaround for a problem which could prevent Net::SMTP->auth() method from working with a server offering DIGEST-MD5 mechanism with qops of "auth-int" or "auth-conf".
This is because Net::SMTP does not support encoding/decoding
traffic over a socket while Authen::SASL supports it and accepts
mechanisms with not only authentication but also encryption.
When the server offers DIGEST-MD5 with "auth-int" or "auth-conf",
the authentication itself succeeds with code 235 but the next
command from the client will fail because Net::SMTP tries to
send it in cleartext.

This breaks DIGEST-MD5 support in Net::SMTP.  To fix this,
this patch disables "auth-int" and "auth-conf" by forcibly
setting maximum SSF on the client side as zero.  By doing this,
the client accepts only "auth" qop in the negotiation even
if the server offers "auth-int" or "auth-conf" in addition to it.

Note that this patch is sub-optimal.  To fix this properly,
Net::SMTP should support READ and WRITE methods provided
by Authen::SASL though it requires major rewrite.