Add support for encode()/decode() methods in Net::Cmd
Some mechanisms in Authen::SASL offer additional security layers and
encode()/decode() methods for them.  The I/O routines in Net::Cmd now use
it if they are available.  Currently only Net::SMTP defines
protocol-specific methods.

In addition to that, EHLO is now issued just after a successful
authentication with an additional security layer as described in
Sec. 4 in RFC 4954.