Do no return $fqdn if domain starts with .
The current logic to get domain name if the hostname does not contain a period but domain name does has one problem. It does not check if the domain name starts with a . or not. In my case, the domain name was simple a period (".") as per this library, and fqdn was `Hostname..`, which was immediately rejected by my email server. So, lets avoid these double dots.